fix transaction value overflow

+++ web3j-app/app/src/main/java/me/bing/web3j/app/mapper/model/Transaction.java
@@ -1,6 +1,5 @@
 package me.bing.web3j.app.mapper.model;
 
-import java.math.BigInteger;
 import java.util.Date;
 
 /**
@@ -13,7 +12,7 @@ public class Transaction {
     private long blockNumber;
     private String from;
     private String to;
-    private BigInteger value;
+    private long value;
     private int status;
     private Date timestamp;
     private int nonce;
@@ -88,11 +87,11 @@ public class Transaction {
         this.to = to;
     }
 
-    public BigInteger getValue() {
+    public long getValue() {
         return value;
     }
 
-    public void setValue(BigInteger value) {
+    public void setValue(long value) {
         this.value = value;
     }
 
@@ -175,7 +174,7 @@ public class Transaction {
         private long blockNumber;
         private String from;
         private String to;
-        private BigInteger value;
+        private long value;
         private int status;
         private Date timestamp;
         private int nonce;
@@ -214,7 +213,7 @@ public class Transaction {
             return this;
         }
 
-        public Builder value(BigInteger val) {
+        public Builder value(long val) {
             value = val;
             return this;
         }
